// src/shared/types/generated/contentful.d.ts
export interface IAlphabetFields {
  /** name */
  name: string;
}
// ...
export interface ICategoryFields {
  /** name */
  name?: string | undefined;
}

export interface ICategory extends Entry<ICategoryFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "category";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IDetailedCategoryFields {
  /** name */
  name?: string | undefined;
}
// ...
export interface IPopularProductFields {
  /** product */
  product?: string | undefined;

  /** skintone */
  skintone?: string | undefined;

  /** undertone */
  undertone?: string | undefined;

  /** category */
  category?: string | undefined;

  /** price */
  price?: string | undefined;

  /** shade_description */
  shade_description?: string | undefined;

  /** link */
  link?: string | undefined;

  /** img_url */
  img_url?: string | undefined;

  /** cruelty_free */
  cruelty_free?: boolean | undefined;

  /** paraben_free */
  paraben_free?: boolean | undefined;

  /** vegan */
  vegan?: boolean | undefined;

  /** gluten_free */
  gluten_free?: boolean | undefined;

  /** oil_free */
  oil_free?: boolean | undefined;

  /** clean */
  clean?: boolean | undefined;

  /** bipoc_founded */
  bipoc_founded?: boolean | undefined;

  /** spf */
  spf?: boolean | undefined;
}
// ...
export interface IPreferenceFields {
  /** name */
  name?: string | undefined;
}

export interface IPreference extends Entry<IPreferenceFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "preference";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPriceFields {
  /** name */
  name: string;
}

export interface IPrice extends Entry<IPriceFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "price";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IProductFields {
  /** category */
  category?: string | undefined;

  /** tutorials */
  tutorials?: number | undefined;

  /** full_product */
  full_product?: string | undefined;

  /** shade */
  shade?: string | undefined;

  /** shade_description */
  shade_description?: string | undefined;

  /** skintone */
  skintone?: string | undefined;

  /** skintone_original */
  skintone_original?: string | undefined;

  /** undertone */
  undertone?: string | undefined;

  /** undertone_original */
  undertone_original?: string | undefined;

  /** price */
  price?: string | undefined;

  /** product_url */
  product_url?: string | undefined;

  /** img_url */
  img_url?: string | undefined;

  /** cruelty_free */
  cruelty_free?: boolean | undefined;

  /** paraben_free */
  paraben_free?: boolean | undefined;

  /** vegan */
  vegan?: boolean | undefined;

  /** gluten_free */
  gluten_free?: boolean | undefined;

  /** oil_free */
  oil_free?: boolean | undefined;

  /** talc_free */
  talc_free?: boolean | undefined;

  /** clean */
  clean?: boolean | undefined;

  /** bipoc_founded */
  bipoc_founded?: boolean | undefined;

  /** spf */
  spf?: number | undefined;

  /** brand */
  brand?: string | undefined;

  /** product */
  product?: string | undefined;
}

export interface IProduct extends Entry<IProductFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "product";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IRecommendationProductFields {
  /** link */
  link?: string | undefined;

  /** creator */
  creator?: string | undefined;

  /** pro */
  pro?: boolean | undefined;

  /** skintone */
  skintone?: string | undefined;

  /** skintone_original */
  skintone_original?: string | undefined;

  /** undertone */
  undertone?: string | undefined;

  /** undertone_original */
  undertone_original?: string | undefined;

  /** foundation_1 */
  foundation_1?: string | undefined;

  /** foundation_2 */
  foundation_2?: string | undefined;

  /** foundation_3 */
  foundation_3?: string | undefined;

  /** foundation_4 */
  foundation_4?: string | undefined;

  /** adjuster_1 */
  adjuster_1?: string | undefined;

  /** adjuster_2 */
  adjuster_2?: string | undefined;

  /** concealer_1 */
  concealer_1?: string | undefined;

  /** concealer_2 */
  concealer_2?: string | undefined;

  /** concealer_3 */
  concealer_3?: string | undefined;

  /** concealer_4 */
  concealer_4?: string | undefined;

  /** concealer_5 */
  concealer_5?: string | undefined;

  /** blush_1 */
  blush_1?: string | undefined;

  /** blush_2 */
  blush_2?: string | undefined;

  /** blush_3 */
  blush_3?: string | undefined;

  /** blush_4 */
  blush_4?: string | undefined;

  /** bronzercontour_1 */
  bronzercontour_1?: string | undefined;

  /** bronzercontour_2 */
  bronzercontour_2?: string | undefined;

  /** bronzercontour_3 */
  bronzercontour_3?: string | undefined;

  /** bronzercontour_4 */
  bronzercontour_4?: string | undefined;

  /** bronzercontour_5 */
  bronzercontour_5?: string | undefined;

  /** powder_1 */
  powder_1?: string | undefined;

  /** powder_2 */
  powder_2?: string | undefined;

  /** powder_3 */
  powder_3?: string | undefined;

  /** powder_4 */
  powder_4?: string | undefined;

  /** powder_5 */
  powder_5?: string | undefined;

  /** powder_6 */
  powder_6?: string | undefined;

  /** highlighter_1 */
  highlighter_1?: string | undefined;

  /** highlighter_2 */
  highlighter_2?: string | undefined;

  /** highlighter_3 */
  highlighter_3?: string | undefined;

  /** lipliner_1 */
  lipliner_1?: string | undefined;

  /** lipliner_2 */
  lipliner_2?: string | undefined;

  /** lipliner_3 */
  lipliner_3?: string | undefined;

  /** lipstick_1 */
  lipstick_1?: string | undefined;

  /** lipstick_2 */
  lipstick_2?: string | undefined;

  /** lipstick_3 */
  lipstick_3?: string | undefined;

  /** lipstick_4 */
  lipstick_4?: string | undefined;

  /** lipgloss_1 */
  lipgloss_1?: string | undefined;

  /** lipgloss_2 */
  lipgloss_2?: string | undefined;

  /** lipgloss_3 */
  lipgloss_3?: string | undefined;

  /** celebrity */
  celebrity?: string | undefined;

  /** post */
  post?: string | undefined;
}
// ...
export interface ISkintoneFields {
  /** name */
  name?: string | undefined;
}

export interface ISkintone extends Entry<ISkintoneFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "skintone";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ISpfFields {
  /** name */
  name?: string | undefined;
}

export interface ISpf extends Entry<ISpfFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "spf";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IUndertoneFields {
  /** name */
  name?: string | undefined;
}